* MB-29598, MB-29481: Count initial checkpoint items memory usage
 * MB-29481: Add CheckpointRemoverTest class and test cases
 * docs/Tracing.md: Document defined tracing categories
 * MB-29548: Remove 'FlusherTask' phosphor trace event

+++ engines/ep/tests/module_tests/kv_bucket_test.h
@@ -77,7 +77,7 @@ public:
     /* Flush the given vbucket to disk, so any outstanding dirty items are
      * written (and are clean).
      */
-    void flush_vbucket_to_disk(uint16_t vbid, int expected = 1);
+    void flush_vbucket_to_disk(uint16_t vbid, size_t expected = 1);
 
     /**
      * Flush the given vBucket to disk if the bucket is peristent, otherwise
